@@ -6,10 +6,10 @@
     "require": {
         "php": "^8.1",
         "advoor/nova-editor-js": "dev-image-service",
-        "artesaos/seotools": "^0.20",
+        "artesaos/seotools": "^1.2",
         "barryvdh/laravel-debugbar": "^3.6",
         "barryvdh/laravel-dompdf": "^2.0",
-        "cviebrock/eloquent-sluggable": "^8.0",
+        "cviebrock/eloquent-sluggable": "^9.0",
         "darryldecode/cart": "^4.2",
         "doctrine/dbal": "^2.6",
         "ebess/advanced-nova-media-library": "^4.0",
@@ -22,49 +22,48 @@
         "giggsey/libphonenumber-for-php": "^8.10",
         "google/apiclient": "^2.4",
         "google/google-apiclient-walletobjects": "^1.20220426",
-        "guzzlehttp/guzzle": "^6.3",
+        "guzzlehttp/guzzle": "^7.8",
         "html2text/html2text": "^4.2",
         "irazasyed/telegram-bot-sdk": "^3.3",
         "kris/laravel-form-builder": "^1.40",
-        "laravel/framework": "^8.75",
+        "laravel/framework": "^9.52",
         "laravel/horizon": "^5.7",
         "laravel/nova": "4.23.0",
         "laravel/sanctum": "^2.15",
         "laravel/scout": "^9.3",
         "laravel/telescope": "^4.6",
         "laravel/tinker": "^2.6",
-        "laravelium/sitemap": "^8.0",
-        "laulamanapps/izettle-api": "dev-develop",
-        "league/commonmark": "^1.5",
-        "league/flysystem-aws-s3-v3": "^1.0",
+        "laulamanapps/izettle-api": "dev-master",
+        "league/commonmark": "^2.4",
+        "league/flysystem-aws-s3-v3": "^3.0",
         "league/glide-laravel": "^1.0",
         "maatwebsite/excel": "^3.1",
         "mischiefcollective/colorjizz": "^1.0",
         "mollie/laravel-mollie": "^2.14",
         "picqer/php-barcode-generator": "^2.2",
         "predis/predis": "^1.1",
-        "sentry/sentry-laravel": "^2.5",
+        "sentry/sentry-laravel": "^4.0",
         "sloveniangooner/searchable-select": "^1.3",
         "smalot/pdfparser": "^0.13",
         "spatie/laravel-csp": "^2.4",
         "spatie/laravel-flash": "^1.3",
-        "spatie/laravel-medialibrary": "^8.10",
+        "spatie/laravel-medialibrary": "^9.12",
         "spatie/laravel-model-states": "^2.4",
         "spatie/laravel-permission": "^5.0",
         "spatie/laravel-view-models": "^1.2",
         "ssnepenthe/color-utils": "^0.4.2",
-        "symfony/process": "^5.4",
+        "symfony/process": "^6.3",
         "symfony/yaml": "^5.0",
         "teamtnt/laravel-scout-tntsearch-driver": "^11.5",
         "whitecube/nova-flexible-content": "^1.0"
     },
     "require-dev": {
         "barryvdh/laravel-ide-helper": "^2.10",
-        "facade/ignition": "^2.17",
+        "spatie/laravel-ignition": "^1.6",
         "fakerphp/faker": "^1.17",
         "friendsofphp/php-cs-fixer": "^3.0",
         "mockery/mockery": "^1.4",
-        "nunomaduro/collision": "^5.10",
+        "nunomaduro/collision": "^6.1",
         "php-parallel-lint/php-console-highlighter": "^0.5.0",
         "php-parallel-lint/php-parallel-lint": "^1.2",
         "phpunit/phpunit": "^9.3",
