@@ -9,8 +9,7 @@ use Illuminate\Http\Response as HttpResponse;
 use Illuminate\Support\Facades\Config;
 use Illuminate\Support\Facades\Response;
 use Illuminate\Support\Facades\Storage;
-use League\Flysystem\FileNotFoundException;
-use League\Glide\Filesystem\FileNotFoundException as GlideFileNotFoundException;
+use League\Glide\Filesystem\FileNotFoundException;
 use League\Glide\Responses\LaravelResponseFactory;
 use League\Glide\ServerFactory;
 use Symfony\Component\HttpFoundation\StreamedResponse;
@@ -40,7 +39,7 @@ class ImageController extends Controller
 
         try {
             return $server->getImageResponse($path, $request->all());
-        } catch (FileNotFoundException | GlideFileNotFoundException) {
+        } catch (FileNotFoundException $e) {
             return Response::noContent(HttpResponse::HTTP_NOT_FOUND);
         }
     }
